replaced unsigned ints with size t in multiple functions

+++ libft_h/libft.a/src/ft_bzero.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 // UNCLEAR IF DONE, REVISIT LATER
-void	ft_bzero(void *s, unsigned int n)
+void	ft_bzero(void *s, size_t n)
 {
 	if (n == 0)
 		return ;
