+++ libft_h/libft.a/src/ft_strncmp.c
@@ -12,9 +12,9 @@
 
 #include "libft.h"
 
-int	ft_strncmp(const char *s1, const char *s2, unsigned int n)
+int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (((s1[i] || s2[i]) != '\0') && i < n)
