+++ libft_h/libft.a/src/ft_strnstr.c
@@ -12,10 +12,10 @@
 
 #include "libft.h"
 
-char	*ft_strnstr(const char *big, const char *little, unsigned int n)
+char	*ft_strnstr(const char *big, const char *little, size_t n)
 {
-	unsigned int	i;
-	unsigned int	j;
+	size_t	i;
+	size_t	j;
 
 	i = 0;
 	if (*little == '\0')
